Rework Dockerfile directive concatenation
This commit slightly shifts the concerns of the provider / function
relationship. A provider is now responsible for implementing a
dockerfile method, which returns the ENTIRE LINE of the Dockerfile, not
the post-directive content.

This change facilitates including lines that don't start with CMD in the
Dockerfile. It seems like the easiest (read: most boring) approach, but
I think it is best for now. Maybe we'll see some dynamic instance_eval
fun in the future.